+++ CHANGELOG.md
@@ -7,6 +7,7 @@ More infos about this file : http://keepachangelog.com/
 ## [Unreleased] - no_due_date
 
 - **DistanceApi#distances method should also accept `Array`s**
+- **Some small optimisations**
 
 ## [v1.0.0-RC1] - 2018.05.14